refactor: split compound simulator UI
