+++ apps/web/src/components/charts/compound-simulator/use-compound-calculator.ts
@@ -6,9 +6,59 @@ import {
 } from "./calculate-compound";
 
 export function useCompoundCalculator(input: CompoundCalculatorInput): YearlyProjection[] {
+  const {
+    initialAmount,
+    monthlyContribution,
+    annualReturnRate,
+    contributionYears,
+    withdrawalStartYear,
+    withdrawalYears,
+    taxFree,
+    monthlyWithdrawal,
+    annualWithdrawalRate,
+    expenseRatio,
+    inflationRate,
+    inflationAdjustedWithdrawal,
+    monthlyPensionIncome,
+    pensionStartYear,
+    monthlyOtherIncome,
+  } = input;
+
   return useMemo(
-    () => calculateCompound(input),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    Object.values(input),
+    () =>
+      calculateCompound({
+        initialAmount,
+        monthlyContribution,
+        annualReturnRate,
+        contributionYears,
+        withdrawalStartYear,
+        withdrawalYears,
+        taxFree,
+        monthlyWithdrawal,
+        annualWithdrawalRate,
+        expenseRatio,
+        inflationRate,
+        inflationAdjustedWithdrawal,
+        monthlyPensionIncome,
+        pensionStartYear,
+        monthlyOtherIncome,
+      }),
+    [
+      initialAmount,
+      monthlyContribution,
+      annualReturnRate,
+      contributionYears,
+      withdrawalStartYear,
+      withdrawalYears,
+      taxFree,
+      monthlyWithdrawal,
+      annualWithdrawalRate,
+      expenseRatio,
+      inflationRate,
+      inflationAdjustedWithdrawal,
+      monthlyPensionIncome,
+      pensionStartYear,
+      monthlyOtherIncome,
+    ],
   );
 }
